Improve update playbook
Move conditonal include of update playbook further up in the main playbook. This makes later tasks, like modifying the git user account, make more sense because they don't exist until GitLab is installed.

Add "UPDATE | " to all tasks in the update playbook so they are easy to distinguish when running.

Change GitLab reconfigure and restart tasks in the update playbook to handlers. This prevents multiple executions of gitlab-ctl reconfigure if the gitlab.rb file is updated at the same time GitLab is updated (or installed initially).

Add handler to wait for unicorn to come back up.