Add handler for new gitlab-ctl task
Notify upgrade gitlab when the package gets updated. According to the GitLab
documentation, this is run automatically by the package after installation with
versions 7.10 and later. This didn't always result in a clean upgrade in my testing,
so I am manually calling the task with a handle instead of relying on the package
to do it.