Remove idempotence test
Change to the apt module now reports a change when the cache is updated, creating a catch 22 with the installation of the package. If the cache timeout is too high, GitLab fails to install since the cache isn't updated. If it's set too low, it updates every time and reports a change.